requirements.txt
ultralytics
opencv-python
keyboard
supabase
python-dotenv
requests
flask
flask-cors
pyzmq